requirements.txt
flask
pillow
gunicorn
segno
